feat(integration): changing cert location
