+++ .github/workflows/push_pr.yml
@@ -77,10 +77,9 @@ jobs:
         run: |
           .\build\windows\unit_tests.ps1
 
-  e2eTests:
+  integrationTests:
     runs-on: ubuntu-20.04
     strategy:
-      max-parallel: 7 # len(k8sVersion) is a good number to have here
       matrix:
         # Latest patch version can be found in https://github.com/kubernetes/sig-release/blob/master/releases/patch-releases.md
         # Some versions might not be available yet in https://storage.googleapis.com/kubernetes-release/release/v1.X.Y/bin/linux/amd64/kubelet
@@ -98,5 +97,6 @@ jobs:
           minikube version: v1.20.0
           kubernetes version: ${{ matrix.k8sVersion }}
       - name: Run Integration Tests
+        env:
+          MINIKUBE_HOME: '/home/runner/work/_temp'
         run: make test-integration
-
